requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
deepgram-sdk
openai
aiofiles
mistralai